Move update_wiki into its own file
