+++ sympy_bot/webapp.py
@@ -2,10 +2,7 @@ import textwrap
 import datetime
 import os
 import base64
-import urllib
-import sys
-import shlex
-from subprocess import run as subprocess_run, CalledProcessError, PIPE
+from subprocess import CalledProcessError
 
 from aiohttp import web, ClientSession
 
@@ -14,6 +11,7 @@ from gidgethub.aiohttp import GitHubAPI
 
 from .changelog import (get_changelog, update_release_notes, VERSION_RE,
                         get_release_notes_filename)
+from .update_wiki import update_wiki
 
 router = routing.Router()
 
@@ -265,68 +263,3 @@ The error message was: {message}
         description='There was an error updating the release notes on the wiki.',
         context='sympy-bot/release-notes',
     ))
-
-# Modified from doctr.travis.run_command_hiding_token
-def run(args, shell=False, check=True):
-    token = os.environ.get("GH_AUTH").encode('utf-8')
-
-    if not shell:
-        command = ' '.join(map(shlex.quote, args))
-    else:
-        command = args
-
-    command = command.replace(token.decode('utf-8'), '~'*len(token))
-    print(command)
-    sys.stdout.flush()
-
-    if token:
-        stdout = stderr = PIPE
-    else:
-        stdout = stderr = None
-    p = subprocess_run(args, stdout=stdout, stderr=stderr, shell=shell, check=check)
-    if token:
-        # XXX: Do this in a way that is streaming
-        out, err = p.stdout, p.stderr
-        out = out.replace(token, b"~"*len(token))
-        err = err.replace(token, b"~"*len(token))
-        if out:
-            print(out.decode('utf-8'))
-        if err:
-            print(err.decode('utf-8'), file=sys.stderr)
-    sys.stdout.flush()
-    sys.stderr.flush()
-    return p.returncode
-
-def update_wiki(*, wiki_url, release_notes_file, changelogs, pr_number,
-                authors):
-    run(['git', 'config', '--global', 'user.email', "sympy+bot@sympy.org"])
-    run(['git', 'config', '--global', 'user.name', "SymPy Bot"])
-
-    run(['git', 'clone', wiki_url, '--depth', '1'], check=True)
-    _, wiki = wiki_url.rsplit('/', 1)
-    os.chdir(wiki)
-
-    with open(release_notes_file, 'r') as f:
-        rel_notes_txt = f.read()
-
-    try:
-        new_rel_notes_txt = update_release_notes(rel_notes_txt=rel_notes_txt,
-        changelogs=changelogs, pr_number=pr_number, authors=authors)
-    except Exception as e:
-        raise RuntimeError(str(e)) from e
-
-    with open(release_notes_file, 'w') as f:
-        f.write(new_rel_notes_txt)
-
-    run(['git', 'diff'], check=True)
-    run(['git', 'add', release_notes_file], check=True)
-
-    message = f"Update {release_notes_file} from PR #{pr_number}"
-    run(['git', 'commit', '-m', message], check=True)
-
-    parsed_url = list(urllib.parse.urlparse(wiki_url))
-    parsed_url[1] = os.environ.get("GH_AUTH") + '@' + parsed_url[1]
-    auth_url = urllib.parse.urlunparse(parsed_url)
-
-    # TODO: Use a deploy key to do this
-    run(['git', 'push', auth_url, 'master'], check=True)
